std::hash does not cryptographically hash values, it's used for hash together values for maps... msvc creates random values but gcc does not. program now works on gcc linux

+++ SandalBotV2/Board.cpp
@@ -161,7 +161,7 @@ namespace SandalBot {
 			BitBoardUtility::printBB(diagonalPieces);
 			BitBoardUtility::printBB(knights);
 			BitBoardUtility::printBB(pawns);
-			cout << stateHistory.getSecondLast().zobristHash << endl;
+			cout << static_cast<uint64_t>(stateHistory.getSecondLast().zobristHash) << endl;
 			throw exception();
 		}
